requirements.txt
# MCP Server - Istedlal AI Agents
# Core
mcp[cli]>=1.0.0
python-dotenv>=1.0.0

# Database (Phase 2 - pgvector)
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Vector alternatives (optional - install if using VECTOR_PROVIDER=chromadb)
# chromadb>=0.4.0

# Async support
# asyncpg>=0.29.0

# Utilities
pydantic>=2.0.0
